storage-mysql/requirements.txt
connexion[flask,swagger-ui,uvicorn]>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.0
pymysql>=1.1.0
SQLAlchemy>=2.0.0
PyYAML>=6.0
kafka-python>=2.0.2
